task 11 completed

+++ restaurants/views.py
@@ -37,7 +37,11 @@ def signin(request):
         "form":form
     }
     return render(request, 'signin.html', context)
-
+def access(request):
+    context = {
+        "message":"you are not allowed"
+    }
+    return render(request, 'access.html', context)
 def signout(request):
     logout(request)
     return redirect("signin")
@@ -59,6 +63,8 @@ def restaurant_detail(request, restaurant_id):
     return render(request, 'detail.html', context)
 
 def restaurant_create(request):
+    if request.user.is_anonymous:
+        return redirect('signin')
     form = RestaurantForm()
     if request.method == "POST":
         form = RestaurantForm(request.POST, request.FILES)
@@ -75,6 +81,11 @@ def restaurant_create(request):
 def item_create(request, restaurant_id):
     form = ItemForm()
     restaurant = Restaurant.objects.get(id=restaurant_id)
+    if request.user.is_anonymous:
+        return redirect('signin')
+    restaurant_obj = Restaurant.objects.get(id=restaurant_id)
+    if not (request.user.is_staff or request.user == restaurant_obj.owner):
+        return redirect('access')
     if request.method == "POST":
         form = ItemForm(request.POST)
         if form.is_valid():
@@ -89,7 +100,11 @@ def item_create(request, restaurant_id):
     return render(request, 'item_create.html', context)
 
 def restaurant_update(request, restaurant_id):
+    if request.user.is_anonymous:
+        return redirect('signin')
     restaurant_obj = Restaurant.objects.get(id=restaurant_id)
+    if not (request.user.is_staff or request.user == restaurant_obj.owner):
+        return redirect('access')
     form = RestaurantForm(instance=restaurant_obj)
     if request.method == "POST":
         form = RestaurantForm(request.POST, request.FILES, instance=restaurant_obj)
@@ -104,5 +119,7 @@ def restaurant_update(request, restaurant_id):
 
 def restaurant_delete(request, restaurant_id):
     restaurant_obj = Restaurant.objects.get(id=restaurant_id)
+    if not request.user.is_staff:
+        return redirect('access')
     restaurant_obj.delete()
     return redirect('restaurant-list')
